fix(web): phase 6 cleanup — extract filter exports, delete dead analytics view, restyle login
Extracts the still-used FilterSelect/DetailState/urlFilterKeys exports out of
task-run-analytics-view.tsx into web/components/filter-select.tsx and
web/lib/task-run-filters.ts, then deletes the file entirely along with the
dead TaskRunAnalyticsView component and its shadowed, unused local
RunDetailPanel. Refines the login page's card layout to match the design
kit's centered aesthetic while preserving all existing OAuth/CSRF/redirect
logic. Consolidates the Accepted departures note in IMPLEMENTATION.md.
