+++ design/reference/web-dashboard/IMPLEMENTATION.md
@@ -144,9 +144,6 @@ GATE 6: full-suite green + final visual pass + departures documented below.
 
 ## Accepted departures
 
-- `checkPRMerged` in `pr_watcher.go` deletes `claw_prs` rows on both merge and
-  close-without-merge. Consequently, terminal `claw_prs.state`, `merged`, and
-  `merged_at` values cannot be observed through `GET /api/claws/:id/prs`; only
-  open rows (or their absence) are visible. If Phase 5's “Delivery (PR rows w/
-  state)” requires terminal PR state, it must source it from `task_run_prs`,
-  which is retained on merge and close.
+- `checkPRMerged` in `pr_watcher.go` deletes `claw_prs` rows when a PR merges or
+  closes without merging, so `GET /api/claws/:id/prs` cannot expose terminal PR
+  state. Phase 5's Delivery rows source terminal state from retained `task_run_prs`.
