+++ web/app/login/page.tsx
@@ -179,8 +179,8 @@ function LoginForm() {
   const isCallback = !!searchParams.get("code")
 
   return (
-    <div className="flex h-screen bg-background items-center justify-center">
-      <div className="w-full max-w-sm space-y-6 p-8 border border-border rounded-xl bg-card">
+    <div className="flex min-h-screen items-center justify-center bg-background">
+      <div className="w-full max-w-sm space-y-6 rounded-xl border border-border bg-card p-8">
         <div className="space-y-1 text-center">
           <h1 className="text-2xl font-semibold tracking-tight">{appName}</h1>
           <p className="text-sm text-muted-foreground">
@@ -200,7 +200,7 @@ function LoginForm() {
               <button
                 type="button"
                 onClick={handleGitHubSignIn}
-                className="w-full flex items-center justify-center gap-2 rounded-md border border-border bg-background px-3 py-2 text-sm font-medium hover:bg-accent hover:text-accent-foreground"
+                className="flex w-full items-center justify-center gap-2 rounded-md border border-border bg-background px-3 py-2 text-sm font-medium hover:bg-accent hover:text-accent-foreground"
               >
                 <svg viewBox="0 0 16 16" className="h-4 w-4 fill-current" aria-hidden="true">
                   <path d="M8 0C3.58 0 0 3.58 0 8c0 3.54 2.29 6.53 5.47 7.59.4.07.55-.17.55-.38 0-.19-.01-.82-.01-1.49-2.01.37-2.53-.49-2.69-.94-.09-.23-.48-.94-.82-1.13-.28-.15-.68-.52-.01-.53.63-.01 1.08.58 1.23.82.72 1.21 1.87.87 2.33.66.07-.52.28-.87.51-1.07-1.78-.2-3.64-.89-3.64-3.95 0-.87.31-1.59.82-2.15-.08-.2-.36-1.02.08-2.12 0 0 .67-.21 2.2.82.64-.18 1.32-.27 2-.27s1.36.09 2 .27c1.53-1.04 2.2-.82 2.2-.82.44 1.1.16 1.92.08 2.12.51.56.82 1.27.82 2.15 0 3.07-1.87 3.75-3.65 3.95.29.25.54.73.54 1.48 0 1.07-.01 1.93-.01 2.2 0 .21.15.46.55.38A8.013 8.013 0 0 0 16 8c0-4.42-3.58-8-8-8z" />
@@ -250,8 +250,8 @@ function LoginFallback() {
   const { appName } = useBranding()
 
   return (
-    <div className="flex h-screen bg-background items-center justify-center">
-      <div className="w-full max-w-sm space-y-6 p-8 border border-border rounded-xl bg-card">
+    <div className="flex min-h-screen items-center justify-center bg-background">
+      <div className="w-full max-w-sm space-y-6 rounded-xl border border-border bg-card p-8">
         <div className="space-y-1 text-center">
           <h1 className="text-2xl font-semibold tracking-tight">{appName}</h1>
           <p className="text-sm text-muted-foreground">Loading…</p>
